+++ tests/components/portfolio/footer/theme-switch/ThemeSwitch.test.tsx
@@ -1,9 +1,13 @@
 import React, { Dispatch } from 'react';
 import { render } from '@testing-library/react';
-import { useStateMock, setBoolStateMock, setStringStateMock } from '../../mocks/use-state-mock';
+import {
+  useStateMock,
+  setBoolStateMock,
+  setStringStateMock,
+} from '../../../../mocks/use-state-mock';
 import ThemeSwitch, {
   switchId,
-} from '../../../src/components/portfolio/footer/theme-switch/ThemeSwitch';
+} from '../../../../../src/components/portfolio/footer/theme-switch/ThemeSwitch';
 
 jest.mock('react', () => ({
   ...jest.requireActual('react'),
