+++ tests/components/portfolio/layout/Layout.test.tsx
@@ -1,25 +1,42 @@
+import React, { Dispatch } from 'react';
 import { render } from '@testing-library/react';
-import Layout, { titleDivId, contentDivId } from '../../../src/components/portfolio/layout/Layout';
+import Layout, {
+  titleDivId,
+  contentDivId,
+} from '../../../../src/components/portfolio/layout/Layout';
+import { contacts } from '../../../mocks/data-mock';
+import { useStateMock, setBoolStateMock, setStringStateMock } from '../../../mocks/use-state-mock';
 
 describe('Layout component', () => {
   const pageTitles = ['Title', 'Title2'];
-  const componentText = 'This is the component';
-  const component = <div>{componentText}</div>;
   const handlePageChange = () => undefined;
+  const useDarkTheme = true;
+  const setUseDarkTheme = setBoolStateMock;
+  const setTheme = setStringStateMock;
   const useWindowsTheme = false;
+  const componentText = 'This is the component';
+  const children = <div>{componentText}</div>;
 
   describe('title is undefined', () => {
     const pageValue = 0;
 
     beforeEach(() => {
+      jest
+        .spyOn(React, 'useState')
+        .mockImplementation(useStateMock as () => [unknown, Dispatch<unknown>]);
+
       render(
         <Layout
           pageValue={pageValue}
           pageTitles={pageTitles}
           handlePageChange={handlePageChange}
+          contacts={contacts}
+          useDarkTheme={useDarkTheme}
+          setUseDarkTheme={setUseDarkTheme}
+          setTheme={setTheme}
           useWindowsTheme={useWindowsTheme}
         >
-          {component}
+          {children}
         </Layout>
       );
     });
@@ -46,9 +63,13 @@ describe('Layout component', () => {
           pageValue={pageValue}
           pageTitles={pageTitles}
           handlePageChange={handlePageChange}
+          contacts={contacts}
+          useDarkTheme={useDarkTheme}
+          setUseDarkTheme={setUseDarkTheme}
+          setTheme={setTheme}
           useWindowsTheme={useWindowsTheme}
         >
-          {component}
+          {children}
         </Layout>
       );
     });
