+++ tests/components/portfolio/tab-panel/TabPanel.test.tsx
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react';
-import TabPanel from '../../../src/components/portfolio/tab-panel/TabPanel';
+import TabPanel from '../../../../src/components/portfolio/tab-panel/TabPanel';
 
 describe('TabPanel component', () => {
   const index = 0;
