+++ tests/mocks/data-mock.tsx
@@ -57,13 +57,12 @@ export const menuItems = menuItemsRawData.map(
       menuItem.window
         ? new WindowModel(
             menuItem.title,
-            <>This is page content</>,
             menuItem.window.windowId,
             menuItem.window.startBarButtonId,
             menuItem.window.isInternetExplorerWindow,
             menuItem.window.heightPercentage,
             menuItem.window.widthPercentage,
-            menuItem.window.addressBarUrl ?? ''
+            menuItem.window.addressBarUrl
           )
         : undefined
     )
@@ -73,7 +72,6 @@ export const windows = menuItems.map((menuItem) => {
   if (menuItem.window) {
     return new WindowModel(
       menuItem.title,
-      menuItem.window.component,
       menuItem.window.windowId,
       menuItem.window.startBarButtonId,
       menuItem.window.isInternetExplorerWindow,
